### 2018-05-14 #733 @whedon
[PRE REVIEW]: NN-SVG: Publication-Ready Neural Network Architecture Schematics
**Submitting author:** @zfrenchee (<a href="http://orcid.org/0000-0001-8173-2315">Alexander LeNail</a>)
**Repository:** <a href="https://github.com/zfrenchee/NN-SVG" target ="_blank">https://github.com/zfrenchee/NN-SVG</a>
**Version:** 1.0.0
