### 2018-05-14 #733 @zfrenchee
Hello!

I added a paper.md and paper.bib file to the repo. Could you try again?
