### 2018-05-15 #733 @tracykteal
Thanks @zfrenchee. I don't see the paper.md and paper.bib in the repo. Could you check to make sure they got in to https://github.com/zfrenchee/NN-SVG? Those will also help the editors understand more about the scope and purpose of the tool.
