+++ paper.bib
@@ -1,0 +1,25 @@
+@INCOLLECTION{AlexNet,
+	title = {ImageNet Classification with Deep Convolutional Neural Networks},
+	author = {Alex Krizhevsky and Sutskever, Ilya and Hinton, Geoffrey E},
+	booktitle = {Advances in Neural Information Processing Systems 25},
+	year = {2012},
+	url = {http://papers.nips.cc/paper/4824-imagenet-classification-with-deep-convolutional-neural-networks.pdf}
+}
+@INPROCEEDINGS{LeNet,
+	title = {Gradient-based learning applied to document recognition},
+	author = {Yann Lecun and Léon Bottou and Yoshua Bengio and Patrick Haffner},
+	booktitle = {Proceedings of the IEEE},
+	year = {1998},
+	pages = {2278--2324}
+}
+@ARTICLE{d3,
+	title = {D3 Data-Driven Documents},
+	author = {Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
+	journal = {IEEE Transactions on Visualization and Computer Graphics},
+	volume = {17},
+	number = {12},
+	year = {2011},
+	url = {http://dx.doi.org/10.1109/TVCG.2011.185},
+	doi = {10.1109/TVCG.2011.185},
+}
+
