### 2018-05-15 #733 @zfrenchee
Hi Tracy,

Sorry, I'm a moron, and forgot to push. Check again?

I've mostly been modeling this off another repo called BioPandas
<https://github.com/rasbt/biopandas> (which is where I heard about JOSS),
so I'm modeling my submission off that repo. I'm new to JOSS, so let me
know if there are other boxes I should be checking!

Thanks,

--Alex
