### 2018-05-17 #733 @arfon
Hi @zfrenchee - thanks for adding the paper. I had a quick review and it's a little light on detail, statement of need, links to existing research outputs using the tool etc. Could you take another look at the recommended paper structure (http://joss.theoj.org/about#paper_structure) and make some additions.

In addition, could you please take a look at the list of volunteer reviewers above and suggest some potential reviewers.
