+++ paper.md
@@ -18,10 +18,12 @@ bibliography: paper.bib
 
 # Summary
 
-NN-SVG is a tool for creating Neural Network (NN) architecture drawings parametrically rather than by hand. It then provides the ability for those researchers to export those drawings to Scalable Vector Graphics (SVG) files, suitable for inclusion in academic papers or as figures on web pages.
+Illustrations of Neural Network architectures are often time-consuming to produce, and machine learning researchers all too often find themselves constructing these diagrams from scratch by hand.
 
-The tool provides the ability to generate figures of three kinds: classic Fully-Connected Neural Network (FCNN) figures, Convolutional Neural Network (CNN) figures of the sort introduced in [@LeNet], and Deep Neural Network figures following the style introduced in [@AlexNet]. The former two are accomplished using the Data-Driven Documents (D3) javascript library [@d3] and the latter with Three.js.
+NN-SVG is a tool for creating Neural Network (NN) architecture drawings parametrically rather than manually. It then provides the ability to export those drawings to Scalable Vector Graphics (SVG) files, suitable for inclusion in academic papers or as figures on web pages.
 
-We hope this tool will save machine learning researchers time which would otherwise be spent drawing figures by hand, and we hope this software might also serve as a pedagogical tool in some contexts.
+The tool provides the ability to generate figures of three kinds: classic Fully-Connected Neural Network (FCNN) figures, Convolutional Neural Network (CNN) figures of the sort introduced in [@LeNet], and Deep Neural Network figures following the style introduced in [@AlexNet]. The former two are accomplished using the D3 javascript library [@d3] and the latter with the javascript library Three.js. NN-SVG provides the ability to style the figure to the user's liking via many size, color, and layout parameters.
+
+We hope this tool will save machine learning researchers time, which is better invested in reseach, and we hope this software might also serve as a pedagogical tool in some contexts.
 
 # References
