### 2018-05-17 #733 @zfrenchee
Hello Arfon,

I've added a statement of need at the top, and just a little extra detail
in the middle.

I submitted this project because someone asked to cite it and I didn't have
something to point them towards, so I suspect there will be research output
using this project in the future, but there is none as of yet. I anticipate
using it for my own purposes in the future as well.

Some potential reviewers:
terrytangyuan
desilinguist
nilesh-patil
rasbt
mlgill
sssomani
kellieotto

On Thu, May 17, 2018 at 4:19 PM Arfon Smith <notifications@github.com>
wrote:

> Hi @zfrenchee <https://github.com/zfrenchee> - thanks for adding the
> paper. I had a quick review and it's a little light on detail, statement of
> need, links to existing research outputs using the tool etc. Could you take
> another look at the recommended paper structure (
> http://joss.theoj.org/about#paper_structure) and make some additions.
>
> In addition, could you please take a look at the list of volunteer
> reviewers above and suggest some potential reviewers.
>
