### 2018-05-18 #733 @arfon
@terrytangyuan @kellieotto - would you be willing to review this submission for JOSS?
