### 2018-05-19 zfrenchee/NN-SVG#4 @csachs
License?
This looks very promising, possibly as well to dynamically embed NN structures in web pages.

What is the license of the code?
