### 2018-05-19 alexlenail/NN-SVG#4 @alexlenail
Thanks for the issue @csachs! It's MIT LICENSE. I just added it. Would love if you show me how you end up using this!! Glad it's of use
