### 2018-05-21 #733 @arfon
@nilesh-patil - would you be willing to review this submission for JOSS?
