### 2018-05-21 #733 @arfon
> @arfon sure, happy to review it. I haven't reviewed a paper for JOSS before, can you share a link for me to go through the process?

@nilesh-patil - sure thing. I'll add some links on the review issue over in #747
