### 2018-05-21 #747 @zfrenchee
Oh, and one important note about this project, @arfon, @nilesh-patil, @terrytangyuan,
NN-SVG is principally a web-tool, not a package. Although this project was built with extensibility in mind ([and I've already been contacted about using it elsewhere](https://github.com/zfrenchee/NN-SVG/issues/4)) the main purpose of this code is to be the basis for [the site](http://alexlenail.me/NN-SVG/). For that reason, there are no installation instructions / API methods / etc...

So I would put a N/A by those checkmarks.
