### 2018-05-21 ee5aa37
Some updates to the README
