+++ paper.md
@@ -24,6 +24,6 @@ NN-SVG is a tool for creating Neural Network (NN) architecture drawings parametr
 
 The tool provides the ability to generate figures of three kinds: classic Fully-Connected Neural Network (FCNN) figures, Convolutional Neural Network (CNN) figures of the sort introduced in [@LeNet], and Deep Neural Network figures following the style introduced in [@AlexNet]. The former two are accomplished using the D3 javascript library [@d3] and the latter with the javascript library Three.js. NN-SVG provides the ability to style the figure to the user's liking via many size, color, and layout parameters.
 
-We hope this tool will save machine learning researchers time, which is better invested in reseach, and we hope this software might also serve as a pedagogical tool in some contexts.
+We hope this tool will save machine learning researchers time, and we hope this software might also serve as a pedagogical tool in some contexts.
 
 # References
