+++ README.md
@@ -1,9 +1,10 @@
+[NN-SVG](http://alexlenail.me/NN-SVG/)
+======
+
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![status](http://joss.theoj.org/papers/52b511ab107595a805107aa4ad70161d/status.svg)](http://joss.theoj.org/papers/52b511ab107595a805107aa4ad70161d)
 
-
-[NN-SVG](http://alexlenail.me/NN-SVG/)
-======
+<img src="./example.svg">
 
 Illustrations of Neural Network architectures are often time-consuming to produce, and machine learning researchers all too often find themselves constructing these diagrams from scratch by hand.
 
