### 2018-05-22 #747 @arfon
> For that reason, there are no installation instructions / API methods / etc...

Just pushing back on this a little @zfrenchee. While I understand that this might not be wrapped up in a library/module with a clean API, I think it's reasonable to expect that installation instructions for this software (i.e. how to run this site myself) would be provided.
