### 2018-05-22 #747 @zfrenchee
@arfon that's an interesting idea. Then again, they're just HTML pages, so there isn't too much I can think of to explain. But I'll keep thinking about it..
