### 2018-06-15 #747 @arfon
> Thanks for the feedback, and sorry for the delayed reply. After some mulling over, I've come to see that you're quite right, and this tool would be more broadly applicable if I make it into a piece of independent code that can be used more broadly. Unfortunately, that's going to require a fair bit of refactoring and re-architecting, which I probably won't have time to get to in the immediate future. If it's alright with you, I'll just leave this open, and hopefully get to it in the next 1-3 weeks. Sound good?

Yes, this is completely fine. I'll mark this as `paused` for now - please let me know on this thread when you're ready to pick this up again.
