+++ LeNet.js
@@ -1,0 +1,267 @@
+
+/////////////////////////////////////////////////////////////////////////////
+                    ///////    Helper Functions    ///////
+/////////////////////////////////////////////////////////////////////////////
+
+let range = n => [...Array(n).keys()];
+
+let nWise = (n, array) => {
+  iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+  iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+  return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+};
+
+let pairWise = (array) => nWise(2, array);
+
+function flatten(array) {
+  return array.reduce(function (flat, toFlatten) {
+    return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
+  }, []);
+}
+
+function isNumber(n) {
+    return !isNaN(parseFloat(n)) && isFinite(n);
+}
+
+let rand = (min, max) => Math.random() * (max - min) + min;
+
+Array.prototype.last = function() { return this[this.length - 1]; };
+
+
+/////////////////////////////////////////////////////////////////////////////
+                      ///////    Variables    ///////
+/////////////////////////////////////////////////////////////////////////////
+
+var w = window.innerWidth;
+var h = window.innerHeight;
+
+var svg = d3.select("#graph-container").append("svg").attr("xmlns", "http://www.w3.org/2000/svg");
+var g = svg.append("g");
+svg.style("cursor", "move");
+
+var color1 = '#e0e0e0';
+var color2 = '#a0a0a0';
+var borderWidth = 1.0;
+var borderColor = "black";
+var rectOpacity = 0.8;
+var betweenSquares = 8;
+var betweenLayers = [];
+var betweenLayersDefault = 12;
+
+// numberOfSquares, squareWidth, stride.
+var architecture = [];
+var lenet = {};
+var layer_offsets = [];
+var largest_layer_width = 0;
+var showLabels = true;
+
+let textFn = (layer) => (typeof(layer) === "object" ? layer['numberOfSquares']+'@'+layer['squareWidth']+'x'+layer['squareWidth'] : "1x"+layer)
+
+var rect = g.selectAll(".rect");
+var conv = g.selectAll(".conv");
+var link = g.selectAll(".link");
+var poly = g.selectAll(".poly");
+var line = g.selectAll(".line");
+var text = g.selectAll(".text");
+var info = g.selectAll(".info");
+
+/////////////////////////////////////////////////////////////////////////////
+                      ///////    Draw Graph    ///////
+/////////////////////////////////////////////////////////////////////////////
+
+function restart() {
+
+    architecture = $('#architecture').find('input[type="number"]').map((i,el) => $(el).val()).get().filter(input => $.isNumeric(input)).map(s => parseInt(s));
+    architecture = range(architecture.length / 3).map(i => architecture.slice(3*i, 3*i+3)).map((l) => {return {'numberOfSquares':l[0],'squareWidth':l[1],'stride':l[2]}});
+
+    lenet.rects = architecture.map((layer, layer_index) => range(layer['numberOfSquares']).map(rect_index => {return {'id':layer_index+'_'+rect_index,'layer':layer_index,'rect_index':rect_index,'side':layer['squareWidth']}}));
+    lenet.rects = flatten(lenet.rects);
+
+    lenet.convs = architecture.map((layer, layer_index) => Object.assign({'id':'conv_'+layer_index,'layer':layer_index}, layer)); lenet.convs.pop();
+    lenet.convs = lenet.convs.map(conv => Object.assign({'x_rel':rand(0.1, 0.9),'y_rel':rand(0.1, 0.9)}, conv))
+
[…]

[…]
