### 2018-07-13 6859353
tried doing stuff with javascript classes
