+++ fcnn.js
@@ -1,207 +1,196 @@
 
-/////////////////////////////////////////////////////////////////////////////
-                    ///////    Helper Functions    ///////
-/////////////////////////////////////////////////////////////////////////////
+class FCNN {
 
-let range = n => [...Array(n).keys()];
+    /////////////////////////////////////////////////////////////////////////////
+                        ///////    Helpers    ///////
+    /////////////////////////////////////////////////////////////////////////////
 
-let nWise = (n, array) => {
-  iterators = Array(n).fill().map(() => array[Symbol.iterator]());
-  iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
-  return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
-};
+    static range(n) { return [...Array(n).keys()]; }
 
-let pairWise = (array) => nWise(2, array);
+    static nWise(n, array) {
+        var iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+        iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+        return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+    };
 
-function flatten(array) {
-  return array.reduce(function (flat, toFlatten) {
-    return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
-  }, []);
+    static pairWise(array) { return FCNN.nWise(2, array); }
+
+    static flatten(array) {
+        return array.reduce((flat, toFlatten) => (flat.concat(Array.isArray(toFlatten) ? FCNN.flatten(toFlatten) : toFlatten)), []);
+    }
+
+    static randomWeight() { return Math.random() * 2 - 1; }
+
+
+    /////////////////////////////////////////////////////////////////////////////
+                          ///////    Constructor    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
+    constructor() {
+
+        this.w = window.innerWidth;
+        this.h = window.innerHeight;
+
+        this.svg = d3.select("#graph-container").append("svg").attr("xmlns", "http://www.w3.org/2000/svg");
+        this.g = this.svg.append("g");
+        this.svg.style("cursor", "move");
+        this.svg.call(d3.zoom().scaleExtent([1 / 2, 8]).on("zoom", () => this.g.attr("transform", d3.event.transform) ));
+
+        this.edgeWidthProportional = false;
+        this.edgeWidth = 0.5;
+        this.weightedEdgeWidth = d3.scaleLinear().domain([0, 1]).range([0, this.edgeWidth]);
+
+        this.edgeOpacityProportional = false;
+        this.edgeOpacity = 1.0
+        this.weightedEdgeOpacity = d3.scaleLinear().domain([0, 1]).range([0, 1]);
+
+        this.edgeColorProportional = false;
+        this.defaultEdgeColor = "#505050";
+        this.negativeEdgeColor = "#0000ff";
+        this.positiveEdgeColor = "#ff0000";
+        this.weightedEdgeColor = d3.scaleLinear().domain([-1, 0, 1]).range([this.negativeEdgeColor, "white", this.positiveEdgeColor]);
+
+        this.nodeDiameter = 20;
+        this.nodeColor = "#ffffff";
+        this.nodeBorderColor = "#333333";
+
+        this.betweenLayers = 160;
+
+        this.architecture = [8, 12, 8];
+        this.betweenNodesInLayer = [20, 20, 20];
+        this.graph = {};
+        this.layer_offsets = [];
+        this.largest_layer_width = 0;
+        this.nnDirection = 'right'
+        this.showBias = false;
+
+        this.showLabels = true;
[…]
