### 2018-07-14 78c6d6f
A closure model
