### 2018-07-14 3a374fb
LeNet basically done
