+++ LeNet.js
@@ -1,267 +1,310 @@
 
-/////////////////////////////////////////////////////////////////////////////
-                    ///////    Helper Functions    ///////
-/////////////////////////////////////////////////////////////////////////////
+function LeNet() {
 
-let range = n => [...Array(n).keys()];
+    /////////////////////////////////////////////////////////////////////////////
+                        ///////    Helper Functions    ///////
+    /////////////////////////////////////////////////////////////////////////////
 
-let nWise = (n, array) => {
-  iterators = Array(n).fill().map(() => array[Symbol.iterator]());
-  iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
-  return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
-};
+    let range = n => [...Array(n).keys()];
 
-let pairWise = (array) => nWise(2, array);
+    let nWise = (n, array) => {
+      iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+      iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+      return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+    };
 
-function flatten(array) {
-  return array.reduce(function (flat, toFlatten) {
-    return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
-  }, []);
+    let pairWise = (array) => nWise(2, array);
+
+    function flatten(array) {
+      return array.reduce(function (flat, toFlatten) {
+        return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
+      }, []);
+    }
+
+    function isNumber(n) {
+        return !isNaN(parseFloat(n)) && isFinite(n);
+    }
+
+    let rand = (min, max) => Math.random() * (max - min) + min;
+
+    Array.prototype.last = function() { return this[this.length - 1]; };
+
+
+    /////////////////////////////////////////////////////////////////////////////
+                          ///////    Variables    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
+    var w = window.innerWidth;
+    var h = window.innerHeight;
+
+    var svg = d3.select("#graph-container").append("svg").attr("xmlns", "http://www.w3.org/2000/svg");
+    var g = svg.append("g");
+    svg.style("cursor", "move");
+
+    var color1 = '#e0e0e0';
+    var color2 = '#a0a0a0';
+    var borderWidth = 1.0;
+    var borderColor = "black";
+    var rectOpacity = 0.8;
+    var betweenSquares = 8;
+    var betweenLayers = [];
+    var betweenLayersDefault = 12;
+
+    // numberOfSquares, squareWidth, stride.
+    var architecture = [];
+    var lenet = {};
+    var layer_offsets = [];
+    var largest_layer_width = 0;
+    var showLabels = true;
+
+    let textFn = (layer) => (typeof(layer) === "object" ? layer['numberOfSquares']+'@'+layer['squareWidth']+'x'+layer['squareWidth'] : "1x"+layer)
+
+    var rect = g.selectAll(".rect");
+    var conv = g.selectAll(".conv");
+    var link = g.selectAll(".link");
+    var poly = g.selectAll(".poly");
+    var line = g.selectAll(".line");
+    var text = g.selectAll(".text");
+    var info = g.selectAll(".info");
+
+    /////////////////////////////////////////////////////////////////////////////
+                          ///////    Draw Graph    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
[…]
