### 2018-07-15 e1b0ee1
AlexNet works
