+++ AlexNet.js
@@ -1,297 +1,341 @@
 
-// /////////////////////////////////////////////////////////////////////////////
-//                     ///////    Helper Functions    ///////
-// /////////////////////////////////////////////////////////////////////////////
+function AlexNet() {
 
-let range = n => [...Array(n).keys()];
+    // /////////////////////////////////////////////////////////////////////////////
+    //                     ///////    Helper Functions    ///////
+    // /////////////////////////////////////////////////////////////////////////////
 
-let nWise = (n, array) => {
-  iterators = Array(n).fill().map(() => array[Symbol.iterator]());
-  iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
-  return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
-};
+    let range = n => [...Array(n).keys()];
 
-let pairWise = (array) => nWise(2, array);
+    let nWise = (n, array) => {
+      iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+      iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+      return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+    };
 
-Array.prototype.last = function() { return this[this.length - 1]; };
+    let pairWise = (array) => nWise(2, array);
 
-let sum = (arr) => arr.reduce((a,b)=>a+b);
+    Array.prototype.last = function() { return this[this.length - 1]; };
 
-let rand = (min, max) => Math.random() * (max - min) + min;
+    let sum = (arr) => arr.reduce((a,b)=>a+b);
 
+    let rand = (min, max) => Math.random() * (max - min) + min;
 
-// /////////////////////////////////////////////////////////////////////////////
-//                       ///////    Variables    ///////
-// /////////////////////////////////////////////////////////////////////////////
 
-var w = window.innerWidth;
-var h = window.innerHeight;
+    // /////////////////////////////////////////////////////////////////////////////
+    //                       ///////    Variables    ///////
+    // /////////////////////////////////////////////////////////////////////////////
 
-var color1 = '#eeeeee';
-var color2 = '#99ddff';
-var color3 = '#ffbbbb';
+    var w = window.innerWidth;
+    var h = window.innerHeight;
 
-var rectOpacity = 0.4;
-var strideOpacity = 0.4;
+    var color1 = '#eeeeee';
+    var color2 = '#99ddff';
+    var color3 = '#ffbbbb';
 
-var line_material = new THREE.LineBasicMaterial( { 'color':0x000000 } );
-var box_material = new THREE.MeshBasicMaterial( {'color':color1, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':rectOpacity, 'depthWrite':false, 'needsUpdate':true} );
-var conv_material = new THREE.MeshBasicMaterial( {'color':color2, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
-var pyra_material = new THREE.MeshBasicMaterial( {'color':color3, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
+    var rectOpacity = 0.4;
+    var strideOpacity = 0.4;
 
-var architecture = [];
-var architecture2 = [];
-var betweenLayers = 20;
+    var line_material = new THREE.LineBasicMaterial( { 'color':0x000000 } );
[…]
