+++ README.md
@@ -11,7 +11,7 @@ NN-SVG is a tool for creating Neural Network (NN) architecture drawings parametr
 
 The tool provides the ability to generate figures of three kinds: classic Fully-Connected Neural Network (FCNN) figures, Convolutional Neural Network (CNN) figures of the sort introduced in [the LeNet paper](http://yann.lecun.com/exdb/publis/pdf/lecun-01a.pdf), and Deep Neural Network figures following the style introduced in [the AlexNet paper](http://papers.nips.cc/paper/4824-imagenet-classification-with-deep-convolutional-neural-networks.pdf). The former two are accomplished using the [D3 javascript library](https://d3js.org/) and the latter with the javascript library [Three.js](https://threejs.org/). NN-SVG provides the ability to style the figure to the user's liking via many size, color, and layout parameters.
 
-We hope this tool will save machine learning researchers time, and we hope this software might also serve as a pedagogical tool in some contexts.
+I hope this tool will save machine learning researchers time, and I hope this software might also serve as a pedagogical tool in some contexts.
 
 <img src="./example.svg">
 
