### 2018-07-15 #747 @zfrenchee
@arfon @terrytangyuan
I've done the refactoring. I believe it will now be possible to use this tool on other sites.
I probably still should add docs about how to do that. Besides that, what else needs doing? What is the way forwards? Should I add the docs and then we do another round of review?
