### 2018-07-15 e086448
Disentangle drawing logic so this code could be used elsewhere (#7)
* Begin Disentangling

* A closure model

* functional I think

* LeNet basically done

* AlexNet works
