+++ fcnn.js
@@ -1,0 +1,291 @@
+
+function FCNN() {
+
+    /////////////////////////////////////////////////////////////////////////////
+                        ///////    Helper Functions    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
+    let range = n => [...Array(n).keys()];
+
+    let nWise = (n, array) => {
+      iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+      iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+      return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+    };
+
+    let pairWise = (array) => nWise(2, array);
+
+    function flatten(array) {
+      return array.reduce(function (flat, toFlatten) {
+        return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
+      }, []);
+    }
+
+    function isNumber(n) {
+        return !isNaN(parseFloat(n)) && isFinite(n);
+    }
+
+    let randomWeight = () => Math.random() * 2 - 1;
+
+    /////////////////////////////////////////////////////////////////////////////
+                          ///////    Variables    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
+    var w = window.innerWidth;
+    var h = window.innerHeight;
+
+    var svg = d3.select("#graph-container").append("svg").attr("xmlns", "http://www.w3.org/2000/svg");
+    var g = svg.append("g");
+    svg.style("cursor", "move");
+
+    var edgeWidthProportional = false;
+    var edgeWidth = 0.5;
+    var weightedEdgeWidth = d3.scaleLinear().domain([0, 1]).range([0, edgeWidth]);
+
+    var edgeOpacityProportional = false;
+    var edgeOpacity = 1.0
+    var weightedEdgeOpacity = d3.scaleLinear().domain([0, 1]).range([0, 1]);
+
+    var edgeColorProportional = false;
+    var defaultEdgeColor = "#505050";
+    var negativeEdgeColor = "#0000ff";
+    var positiveEdgeColor = "#ff0000";
+    var weightedEdgeColor = d3.scaleLinear().domain([-1, 0, 1]).range([negativeEdgeColor, "white", positiveEdgeColor]);
+
+    var nodeDiameter = 20;
+    var nodeColor = "#ffffff";
+    var nodeBorderColor = "#333333";
+
+    var betweenLayers = 160;
+
+    var architecture = [8, 12, 8];
+    var betweenNodesInLayer = [20, 20, 20];
+    var graph = {};
+    var layer_offsets = [];
+    var largest_layer_width = 0;
+    var nnDirection = 'right'
+    var showBias = false;
+
+    var showLabels = true;
+    let sup_map = {'0': '⁰', '1': '¹', '2': '²', '3': '³', '4': '⁴', '5': '⁵', '6': '⁶', '7': '⁷', '8': '⁸', '9': '⁹'};
+    let sup = (s) => Array.prototype.map.call(s, (d) => (d in sup_map && sup_map[d]) || d).join('');
+
+    let textFn = (layer_index, layer_width) => ((layer_index === 0 ? "Input" : (layer_index === architecture.length-1 ? "Output" : "Hidden")) + " Layer ∈ ℝ" + sup(layer_width.toString()));
+    var nominal_text_size = 12;
+    var textWidth = 70;
+
+    var link = g.selectAll(".link");
+    var node = g.selectAll(".node");
+    var text = g.selectAll(".text");
+
+    /////////////////////////////////////////////////////////////////////////////
+                          ///////    Draw Graph    ///////
+    /////////////////////////////////////////////////////////////////////////////
+
+    function redraw(p_architecture = architecture) {
+
+        architecture = p_architecture;
+
[…]
