### 2018-07-15 55f8f96
Move towards cleaner API
