### 2018-07-15 e9b3334
found lenet's architecture2
