### 2018-07-15 ab5c9ea
clean up helpers
