### 2018-07-15 f35068d
remove isNumber
