+++ LeNet.js
@@ -17,8 +17,6 @@ function LeNet() {
 
     let flatten = (array) => array.reduce((flat, toFlatten) => (flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten)), []);
 
-    let isNumber = (n) => !isNaN(parseFloat(n)) && isFinite(n);
-
     let rand = (min, max) => Math.random() * (max - min) + min;
 
     Array.prototype.last = function() { return this[this.length - 1]; };
