### 2018-07-15 21f1473
Even cleaner
