### 2018-07-15 74c1346
Cleaner API (#8)
* Move towards cleaner API

* 3 methods total

* LeNet

* AlexNet

* found lenet's architecture2

* fcnn.js -> FCNN.js ?

* clean up helpers

* remove isNumber

* Even cleaner
