+++ fcnn.js
@@ -8,24 +8,17 @@ function FCNN() {
     let range = n => [...Array(n).keys()];
 
     let nWise = (n, array) => {
-      iterators = Array(n).fill().map(() => array[Symbol.iterator]());
-      iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
-      return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+        iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+        iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+        return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
     };
 
     let pairWise = (array) => nWise(2, array);
 
-    function flatten(array) {
-      return array.reduce(function (flat, toFlatten) {
-        return flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten);
-      }, []);
-    }
-
-    function isNumber(n) {
-        return !isNaN(parseFloat(n)) && isFinite(n);
-    }
+    let flatten = (array) => array.reduce((flat, toFlatten) => (flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten)), []);
 
     let randomWeight = () => Math.random() * 2 - 1;
+
 
     /////////////////////////////////////////////////////////////////////////////
                           ///////    Variables    ///////
@@ -82,9 +75,13 @@ function FCNN() {
                           ///////    Draw Graph    ///////
     /////////////////////////////////////////////////////////////////////////////
 
-    function redraw(p_architecture = architecture) {
+    function redraw({architecture_=architecture,
+                     showBias_=showBias,
+                     showLabels_=showLabels}={}) {
 
-        architecture = p_architecture;
+        architecture = architecture_;
+        showBias = showBias_;
+        showLabels = showLabels_;
 
         graph.nodes = architecture.map((layer_width, layer_index) => range(layer_width).map(node_index => {return {'id':layer_index+'_'+node_index,'layer':layer_index,'node_index':node_index}}));
         graph.links = pairWise(graph.nodes).map((nodes) => nodes[0].map(left => nodes[1].map(right => {return right.node_index >= 0 ? {'source':left.id,'target':right.id,'weight':randomWeight()} : {} })));
@@ -99,9 +96,6 @@ function FCNN() {
                    .insert("line", ".node")
                    .attr("class", "link")
                    .merge(link);
-        setEdgeWidth();
-        setEdgeOpacity();
-        setEdgeColor();
 
         node = node.data(graph.nodes);
         node.exit().remove();
@@ -113,7 +107,6 @@ function FCNN() {
                    .on("mousedown", set_focus)
                    .on("mouseup", remove_focus)
                    .merge(node);
-        setNodeStyles();
 
         text = text.data(label);
         if (text.empty()) { text = text.enter().append("text").attr("class", "text"); }
@@ -122,11 +115,16 @@ function FCNN() {
                    .style("font-size", nominal_text_size+"px")
                    .merge(text);
 
+        style();
     }
 
-    function redistribute(p_betweenNodesInLayer = betweenNodesInLayer) {
+    function redistribute({betweenNodesInLayer_=betweenNodesInLayer,
+                           betweenLayers_=betweenLayers,
+                           nnDirection_=nnDirection}={}) {
 
-        betweenNodesInLayer = p_betweenNodesInLayer;
+        betweenNodesInLayer = betweenNodesInLayer_;
+        betweenLayers = betweenLayers_;
[…]

[…]
