### 2018-07-15 0f8a844
change name of FCNN.js
