### 2018-07-15 #747 @zfrenchee
@arfon @terrytangyuan
I've written some docs: https://github.com/zfrenchee/NN-SVG/wiki/NN-SVG-API
I believe it makes sense to update the paper slightly, but I also think I'm ready for the next round of review, unless I'm forgetting something?
