### 2018-09-12 #747 @terrytangyuan
Cool. Also, could you address my comments regarding community guidelines?
