### 2018-09-12 #747 @zfrenchee
- @terrytangyuan Could you share an example of what you mean by community guidelines? I'm not seeing anything that obviously seems like community guidelines in [BioPandas](https://github.com/rasbt/biopandas).
- @terrytangyuan I'm not sure it makes sense for a visual tool to provide automated tests?
- @terrytangyuan the wiki documents all the API methods for each architecture. Each architecture only has 2-3 methods, which might have confused you?
