### 2018-09-12 #747 @terrytangyuan
There is no descriptions for each parameters in the public methods.

You can find example here: http://rasbt.github.io/biopandas/contributing/#how-to-contribute
It's located in the README.md: https://github.com/rasbt/biopandas#links
You can also provide a CONTRIBUTING.md file in the repo.

Regarding automated tests for visual tool, @arfon any suggestions?
