### 2018-09-12 #747 @arfon
> Regarding automated tests for visual tool, @arfon any suggestions?

Are the visualizations in HTML? If so, something like Selenium might be a good fit?
