### 2018-09-12 #747 @zfrenchee
@arfon that's a good point, but I think that would be overkill for this project. From the little frontend testing I have written, I can tell this would require 2-3x the code I've already written to write robust UI tests. That doesn't seem called for by this application. What do you think?
