### 2018-09-13 #747 @arfon
@zfrenchee - yeah, I hear you. Automated tests aren't a firm requirement for JOSS but if they're not present then something similar (documented test procedures with example data etc.) needs to be provided instead. Basically the reviewers need to be able to verify the functionality of the software.
