### 2018-09-14 #747 @zfrenchee
@arfon that makes a lot of sense. Let me see what I can rustle up in the way of something like tests.
