+++ UNet.js
@@ -1,0 +1,345 @@
+
+function UNet() {
+
+    // /////////////////////////////////////////////////////////////////////////////
+    //                     ///////    Helper Functions    ///////
+    // /////////////////////////////////////////////////////////////////////////////
+
+    let nWise = (n, array) => {
+      iterators = Array(n).fill().map(() => array[Symbol.iterator]());
+      iterators.forEach((it, index) => Array(index).fill().forEach(() => it.next()));
+      return Array(array.length - n + 1).fill().map(() => (iterators.map(it => it.next().value)));
+    };
+
+    let pairWise = (array) => nWise(2, array);
+
+    let sum = (arr) => arr.reduce((a,b)=>a+b);
+
+    let rand = (min, max) => Math.random() * (max - min) + min;
+
+    Array.prototype.last = function() { return this[this.length - 1]; };
+
+
+    // /////////////////////////////////////////////////////////////////////////////
+    //                       ///////    Variables    ///////
+    // /////////////////////////////////////////////////////////////////////////////
+
+    var w = window.innerWidth;
+    var h = window.innerHeight;
+
+    var color1 = '#eeeeee';
+    var color2 = '#99ddff';
+    var color3 = '#ffbbbb';
+
+    var rectOpacity = 0.4;
+    var strideOpacity = 0.4;
+
+    var line_material = new THREE.LineBasicMaterial( { 'color':0x000000 } );
+    var box_material = new THREE.MeshBasicMaterial( {'color':color1, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':rectOpacity, 'depthWrite':false, 'needsUpdate':true} );
+    var conv_material = new THREE.MeshBasicMaterial( {'color':color2, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
+    var pyra_material = new THREE.MeshBasicMaterial( {'color':color3, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
+
+    var architecture = [];
+    var architecture2 = [];
+    var betweenLayers = 20;
+
+    var logDepth = true;
+    var depthScale = 10;
+    var logWidth = true;
+    var widthScale = 10;
+    var logConvSize = false;
+    var convScale = 1;
+
+    var showDims = false;
+
+    let depthFn = (depth) => logDepth ? (Math.log(depth) * depthScale) : (depth * depthScale);
+    let widthFn = (width) => logWidth ? (Math.log(width) * widthScale) : (width * widthScale);
+    let convFn = (conv) => logConvSize ? (Math.log(conv) * convScale) : (conv * convScale);
+
+    function wh(layer) { return widthFn(layer['widthAndHeight']); }
+
+    var layers = new THREE.Group();
+    var convs = new THREE.Group();
+    var pyramids = new THREE.Group();
+    var sprites = new THREE.Group();
+
+
+    var scene = new THREE.Scene();
+    scene.background = new THREE.Color( 0xffffff );
+
+    // var camera = new THREE.PerspectiveCamera( 75, window.innerWidth/window.innerHeight, 0.1, 100000 );
+    var camera = new THREE.OrthographicCamera( w / - 2, w / 2, h / 2, h / - 2, -10000000, 10000000 );
+    camera.position.set(-219, 92, 84);
+
+    var renderer;
+    var rendererType = 'webgl';
+
+    var controls;
+
+
+    // /////////////////////////////////////////////////////////////////////////////
+    //                       ///////    Draw Graph    ///////
+    // /////////////////////////////////////////////////////////////////////////////
+
[…]
