+++ UNet.js
@@ -27,17 +27,14 @@ function UNet() {
     var w = window.innerWidth;
     var h = window.innerHeight;
 
-    var color1 = '#eeeeee';
-    var color2 = '#99ddff';
-    var color3 = '#ffbbbb';
-
-    var rectOpacity = 0.4;
-    var strideOpacity = 0.4;
+    var clr_vol = '#eeeeee';
+    var clr_conv = '#99ddff';
+    var clr_upconv = '#ffbbbb';
+    var clr_pool = '#ffbbbb';
+    var clr_conn = '#ffbbbb';
 
     var line_material = new THREE.LineBasicMaterial( { 'color':0x000000 } );
-    var box_material = new THREE.MeshBasicMaterial( {'color':color1, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':rectOpacity, 'depthWrite':false, 'needsUpdate':true} );
-    var conv_material = new THREE.MeshBasicMaterial( {'color':color2, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
-    var pyra_material = new THREE.MeshBasicMaterial( {'color':color3, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':strideOpacity, 'depthWrite':false, 'needsUpdate':true} );
+    var vol_material = new THREE.MeshBasicMaterial( {'color':clr_vol, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':1, 'depthWrite':false, 'needsUpdate':true} );
 
     var architecture = [];
     var architecture2 = [];
@@ -47,20 +44,15 @@ function UNet() {
     var depthScale = 10;
     var logWidth = true;
     var widthScale = 10;
-    var logConvSize = false;
-    var convScale = 1;
 
     var showDims = false;
 
     let depthFn = (depth) => logDepth ? (Math.log(depth) * depthScale) : (depth * depthScale);
     let widthFn = (width) => logWidth ? (Math.log(width) * widthScale) : (width * widthScale);
-    let convFn = (conv) => logConvSize ? (Math.log(conv) * convScale) : (conv * convScale);
 
     function wh(layer) { return widthFn(layer['widthAndHeight']); }
 
     var layers = new THREE.Group();
-    var convs = new THREE.Group();
-    var pyramids = new THREE.Group();
     var sprites = new THREE.Group();
 
 
@@ -118,8 +110,6 @@ function UNet() {
                      depthScale_=depthScale,
                      logWidth_=logWidth,
                      widthScale_=widthScale,
-                     logConvSize_=logConvSize,
-                     convScale_=convScale,
                      showDims_=showDims}={}) {
 
         architecture = architecture_;
@@ -129,8 +119,6 @@ function UNet() {
         depthScale = depthScale_;
         logWidth = logWidth_;
         widthScale = widthScale_;
-        logConvSize = logConvSize_;
-        convScale = convScale_;
         showDims = showDims_;
 
         clearThree(scene);
@@ -143,7 +131,7 @@ function UNet() {
 
             // Layer
             layer_geometry = new THREE.BoxGeometry( wh(layer), wh(layer), depthFn(layer['depth']) );
-            layer_object = new THREE.Mesh( layer_geometry, box_material );
+            layer_object = new THREE.Mesh( layer_geometry, vol_material );
             layer_object.position.set(0, 0, layer_offsets[index]);
             layers.add( layer_object );
 
@@ -151,44 +139,6 @@ function UNet() {
             layer_edges_object = new THREE.LineSegments( layer_edges_geometry, line_material );
             layer_edges_object.position.set(0, 0, layer_offsets[index]);
             layers.add( layer_edges_object );
-
-            if (index < architecture.length - 1) {
-
[…]
