+++ UNet.js
@@ -27,17 +27,20 @@ function UNet() {
     var w = window.innerWidth;
     var h = window.innerHeight;
 
-    var clr_vol = '#eeeeee';
-    var clr_conv = '#99ddff';
-    var clr_upconv = '#ffbbbb';
-    var clr_pool = '#ffbbbb';
-    var clr_conn = '#ffbbbb';
+    var clr_vol = '#5a9dd6';
+    var clr_conv = '#7030a0';
+    var clr_upconv = '#ffc000';
+    var clr_pool = '#c10000';
+    var clr_conn = '#6dae47';
+
+    var rectOpacity = 0.4;
+
 
     var line_material = new THREE.LineBasicMaterial( { 'color':0x000000 } );
-    var vol_material = new THREE.MeshBasicMaterial( {'color':clr_vol, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':1, 'depthWrite':false, 'needsUpdate':true} );
+    var vol_material = new THREE.MeshBasicMaterial( {'color':clr_vol, 'side':THREE.DoubleSide, 'transparent':true, 'opacity':rectOpacity, 'depthWrite':false, 'needsUpdate':true} );
 
     var architecture = [];
-    var architecture2 = [];
+    var connections = [];
     var betweenLayers = 20;
 
     var logDepth = true;
@@ -46,6 +49,10 @@ function UNet() {
     var widthScale = 10;
 
     var showDims = false;
+
+    var colors = {"Conv" : clr_conv,
+                "Pooling" : clr_pool,
+                "Up-Conv": clr_upconv}
 
     let depthFn = (depth) => logDepth ? (Math.log(depth) * depthScale) : (depth * depthScale);
     let widthFn = (width) => logWidth ? (Math.log(width) * widthScale) : (width * widthScale);
@@ -104,7 +111,7 @@ function UNet() {
     restartRenderer();
 
     function redraw({architecture_=architecture,
-                     architecture2_=architecture2,
+                     connections_=connections,
                      betweenLayers_=betweenLayers,
                      logDepth_=logDepth,
                      depthScale_=depthScale,
@@ -113,7 +120,7 @@ function UNet() {
                      showDims_=showDims}={}) {
 
         architecture = architecture_;
-        architecture2 = architecture2_;
+        connections = connections_;
         betweenLayers = betweenLayers_;
         logDepth = logDepth_;
         depthScale = depthScale_;
@@ -122,14 +129,17 @@ function UNet() {
         showDims = showDims_;
 
         clearThree(scene);
+        var widths= [];
 
         z_offset = -(sum(architecture.map(layer => depthFn(layer['depth']))) + (betweenLayers * (architecture.length - 1))) / 3;
         layer_offsets = pairWise(architecture).reduce((offsets, layers) => offsets.concat([offsets.last() + depthFn(layers[0]['depth'])/2 + betweenLayers + depthFn(layers[1]['depth'])/2]), [z_offset]);
-        layer_offsets = layer_offsets.concat(architecture2.reduce((offsets, layer) => offsets.concat([offsets.last() + widthFn(2) + betweenLayers]), [layer_offsets.last() + depthFn(architecture.last()['depth'])/2 + betweenLayers + widthFn(2)]));
+        layer_offsets = layer_offsets.concat(connections.reduce((offsets, layer) => offsets.concat([offsets.last() + widthFn(2) + betweenLayers]), [layer_offsets.last() + depthFn(architecture.last()['depth'])/2 + betweenLayers + widthFn(2)]));
 
         architecture.forEach( function( layer, index ) {
 
             // Layer
+            widths.push(wh(layer))
+
[…]
