### 2018-10-04 d3a9ad9
Add vertical leveling to UNet
