### 2018-10-07 #747 @arfon
> @arfon that makes a lot of sense. Let me see what I can rustle up in the way of something like tests.

How are you getting along @zfrenchee?
