### 2018-10-11 d275860
fix a bug when removing layers
