### 2018-12-28 2d4baf4
Should fix #15 & #16
