### 2018-12-28 #747 @zfrenchee
I created a set of minimal examples which demonstrate how to get up and running with this project:

https://github.com/zfrenchee/NN-SVG/wiki/Tests---Minimal-Examples

With those examples, you should be able to draw an NN, and configure it with any of the display parameters listed in the API docs:

https://github.com/zfrenchee/NN-SVG/wiki/NN-SVG-API#iii-style-the-network

I think a reasonable test case is to clone thee repo, copy paste a minimal example, verify that you see an NN diagram, and then configure or style it in various ways using the API.
