### 2019-01-13 #747 @arfon
Thanks @zfrenchee.

@terrytangyuan - could you take a look at the examples @zfrenchee has added here and see if:

> - Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)
