### 2019-01-14 #747 @terrytangyuan
Thanks @zfrenchee for addressing those and thanks @arfon for the ping. That looks fine. Could you address the following comments that I posted earlier?
* There's no community guidelines for how to contribute from other people (how to extend to new NN structure beyond just LeNet, AlexNet, FCNN, etc given how fast this field is evolving).
* Missing installation instructions (this is too trivial @arfon do we need this?)
