### 2019-01-14 #747 @arfon
> - Missing installation instructions (this is too trivial @arfon do we need this?)

I guess since most (all?) of the dependencies here are JavaScript libraries that a pulled from CDNs on the web, perhaps we don't need installation instructions?

That said, it might be good to make clear in the README (or wiki) what the end-user needs to generate these figures.
