### 2019-01-14 dfdb787
Add docs and contributing link to homepage
