### 2019-01-14 #747 @zfrenchee
Hi @arfon, @terrytangyuan

1. "Installation Instructions": The wiki describes the necessary javascripts to include via CDN, implicitly in the [Minimal examples](https://github.com/zfrenchee/NN-SVG/wiki/Tests---Minimal-Examples) and explicitly in the [API docs](https://github.com/zfrenchee/NN-SVG/wiki/NN-SVG-API#i-initialize-the-object)

2. [Contributing Guidelines added to the wiki](https://github.com/zfrenchee/NN-SVG/wiki/Contributing), and linked to in the README.
