### 2019-01-15 #747 @zfrenchee
@arfon sorry to be a bother -- could you point me towards a little more information about this step in the process? I don't see anything on the [JOSS submission instructions page](https://joss.theoj.org/about#author_guidelines), and I'm slightly unsure about what's being asked of me. Am I hearing correctly that I should archive the current code on zenodo? Just downloading a zip of the repo from github and uploading to zenodo would fulfill this?
