### 2019-01-15 #747 @arfon
> Am I hearing correctly that I should archive the current code on zenodo? Just downloading a zip of the repo from github and uploading to zenodo would fulfill this?

Yes, that's correct. If you have made substantial changes to the codebase/docs during review then we also encourage you to make a new release on GitHub. You can learn how to do this by following this guide: https://guides.github.com/activities/citable-code/
