### 2019-01-15 #747 @zfrenchee
Thanks @arfon

See https://zenodo.org/record/2541121#.XD49sM9KjmE and the [releases page of NN-SVG](https://github.com/zfrenchee/NN-SVG/releases)
